aoi_quality_inspection/requirements.txt
# Web框架
Flask==2.3.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.0
Flask-RESTX==1.3.0

# 深度学习
torch==2.0.1
torchvision==0.15.2
onnx==1.14.0
onnxruntime==1.15.1

# 向量搜索
faiss-cpu==1.7.4

# 图像处理
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3

# 数据库
SQLAlchemy==2.0.20

# 文件监控
watchdog==3.0.0

# gRPC
grpcio==1.57.0
grpcio-tools==1.57.0
protobuf==4.24.0

# 配置管理
PyYAML==6.0.1
python-dotenv==1.0.0

# 日志
loguru==0.7.0

# 测试
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0

# 性能测试
locust==2.15.0

# 代码质量
black==23.7.0
flake8==6.1.0
mypy==1.5.0

# 工具
tqdm==4.66.0
requests==2.31.0
